ci tweaks
