+++ .codesandbox/templates/reach-ui/src/index.js
@@ -1,26 +1,12 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import CurrentState from "./CurrentState";
-
+import App from "./App";
 import "./styles.css";
 import "./reach-styles";
 
-function App() {
-  return (
-    <div className="App">
-      <h1>Reach UI Template</h1>
-      <div style={{ margin: "1rem 0" }}>{/* Add your code here. */}</div>
-      <hr />
-      <CurrentState
-        {
-          ...{
-            /* put any props/state here to inspect on the screen */
-          }
-        }
-      />
-    </div>
-  );
-}
+// Open App.tsx to add your code If you would prefer not to write TypeScript,
+// rename App.tsx to App.js, refresh the CodeSandbox browser and your code
+// should still work! We ❤️ CodeSandbox!
 
 const rootElement = document.getElementById("root");
 ReactDOM.render(<App />, rootElement);
